3rd merge

+++ src/main/java/org/zerock/controller/RestaurantController.java
@@ -2,21 +2,23 @@ package org.zerock.controller;
 
 import java.util.List;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpSession;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 import org.zerock.domain.member.MemberVO;
-import org.zerock.domain.restaurant.AddressVO;
+import org.zerock.domain.restaurant.RAddressVO;
 import org.zerock.domain.restaurant.Rcriteria;
 import org.zerock.domain.restaurant.RestaurantVO;
 import org.zerock.domain.restaurant.RpageDTO;
+import org.zerock.service.restaurant.RFileUpService;
 import org.zerock.service.restaurant.RestaurantService;
 
 import lombok.AllArgsConstructor;
@@ -28,6 +30,7 @@ import lombok.extern.log4j.Log4j;
 @Log4j
 public class RestaurantController {
 	private RestaurantService service;
+	private RFileUpService fileUpSvc;
 
 	@GetMapping("/list")
 	public void list(Model model, @ModelAttribute("cri") Rcriteria cri) {
@@ -44,7 +47,7 @@ public class RestaurantController {
 	}
 
 	@PostMapping("/register")
-	public String register(RestaurantVO restaurant, BindingResult result, RedirectAttributes rttr, AddressVO addr, HttpSession session) throws Exception {
+	public String register(RestaurantVO restaurant, MultipartFile file, RedirectAttributes rttr, RAddressVO addr, HttpSession session) throws Exception {
 		// manager == 1 세션 가져오기
 		// User authUser = (User) req.getSession().getAttribute("authUser");
 //		MemberVO user = (MemberVO) session.getAttribute("authUser");
@@ -52,12 +55,22 @@ public class RestaurantController {
 		log.info("*************** m.name ***********" + restaurant.getMname() + "******************************");
 		MemberVO user = (MemberVO) session.getAttribute("authUser");
 		
+		restaurant.setFilename("");
 		if (user.getManager() == 1) {
 			String address = addr.getAddress1() + " " + addr.getAddress2();
 			log.info("**************************" + address + "******************************");
 			restaurant.setRloc(address);
 			log.info("**************************" + restaurant + "******************************");
 			service.register(restaurant);
+			if(file != null) {
+				restaurant.setFilename("restaurant_"+restaurant.getNo()+"_"+file.getOriginalFilename());
+				service.modify(restaurant);
+				try {
+					fileUpSvc.transfer(file, restaurant.getFilename());
+					} catch (Exception e) {
+						e.printStackTrace();
+					}
+			}
 			rttr.addFlashAttribute("result", "success");
 			rttr.addFlashAttribute("message", restaurant.getNo() + "번 글이 등록되었습니다");
 		}
@@ -104,7 +117,7 @@ public class RestaurantController {
 		}
 		RestaurantVO vo = service.read(no);
 		log.info("********* modify get *************" + vo.getRloc() + "*******************");
-
+		log.info("********* modify get *************" + vo.getFilename() + "*******************");
 		model.addAttribute("restaurant", vo);
 		return "/restaurant/modify";
 	}
@@ -116,10 +129,11 @@ public class RestaurantController {
 //			rttr.addFlashAttribute("result", "success");
 //			rttr.addFlashAttribute("message", no + "번 글이 삭제되었습니다");
 //		}
-//		MemberVO user = (MemberVO) session.getAttribute("authUser");
-//		if (user.getManager() == 1) {
+		MemberVO user = (MemberVO) session.getAttribute("authUser");
+		if (user.getManager() == 1) {
 			service.remove(no);
-//		}
+
+			}
 			rttr.addAttribute("pageNo", cri.getPageNo());
 			rttr.addAttribute("amount", cri.getAmount());
 			rttr.addAttribute("type", cri.getType());
@@ -131,24 +145,39 @@ public class RestaurantController {
 	@PostMapping("/modify")
 	// manager 세션 가져오기
 	public String modify(RestaurantVO restaurant, RedirectAttributes rttr, Rcriteria cri, HttpSession session,
-			AddressVO addr) throws Exception {
+			RAddressVO addr, MultipartFile file) throws Exception {
 		String address = addr.getAddress1() + " " + addr.getAddress2();
+		RestaurantVO vo = new RestaurantVO();
+		vo = service.read(restaurant.getNo());
 		log.info(restaurant);
-		log.info("****************    address   **********" + address + "******************************");
-//		MemberVO user = (MemberVO) session.getAttribute("authUser");
-//		if (user.getManager() == 1) {
-
+		log.info("****************    filename   **********" + file.getOriginalFilename() + "******************************");
+		log.info("****************    vo.filename   **********" + vo.getFilename() + "******************************");
+		MemberVO user = (MemberVO) session.getAttribute("authUser");
+		if (user.getManager() == 1) {
 			restaurant.setRloc(address);
-
-			if (service.modify(restaurant)) {
+			restaurant.setFilename(vo.getFilename());
+			service.modify(restaurant);
+			// 파일이 null이 아닐때
+			if(file != null && file.getSize() > 0) {
+				restaurant.setFilename("restaurant_"+restaurant.getNo()+"_"+file.getOriginalFilename());
+				service.modify(restaurant);
 				rttr.addFlashAttribute("result", "success");
 				rttr.addFlashAttribute("message", restaurant.getNo() + "번 글이 수정되었습니다");
-			}
-//		}
+				try {
+					fileUpSvc.transfer(file, restaurant.getFilename());
+					} catch (Exception e) {
+						e.printStackTrace();
+					}
+			  } 
+
+			rttr.addFlashAttribute("result", "success");
+			rttr.addFlashAttribute("message", restaurant.getNo() + "번 글이 수정되었습니다");
+		}
 		rttr.addAttribute("pageNo", cri.getPageNo());
 		rttr.addAttribute("amount", cri.getAmount());
 		rttr.addAttribute("type", cri.getType());
 		rttr.addAttribute("keyword", cri.getKeyword());
+
 
 		return "redirect:/restaurant/list";
 	}
